+++ 12.微信读书助手/导出资料/人类酷刑简史.txt
@@ -1,0 +1,8 @@
+# 作者序
+> 对于可怕的纳粹德国毒气室和曾经风行的电棍，几乎每个有社会常识的人都很熟悉，因而没有收录书中的必要。
+
+
+## 中世纪的酷刑
+> 诵经结束后，一个牧师布置好断头台，然后开始宣读判决，象征性地把犯人交到世俗机构手里，并伪善地请求他们不要以任何方式伤害这些可怜的迷途羔羊。
+
+
