+++ 12.微信读书助手/导出资料/你不可不知的法律常识.txt
@@ -1,0 +1,65 @@
+## 1.酒后出事，同席者承担多大责任？
+> 第二，同席者都喝酒了，但没有相互劝酒，也对喝醉的人及时提醒。一般同席者很难举证证明自己进行了及时提醒，所以法院出于人道主义的目的，可能会判决其承担少量的赔偿，金额一般不会太大。
+第三，同席者都喝酒了，但个别人中途离场，期间也未劝过酒。这种情况下，中途离场的人一般不需要承担责任。
+
+
+## 5.遇到警察查证件，没带身份证怎么办？
+> 按照要求，盘查一般由两名以上民警进行；应当身着警服，未穿警服的，应当出示警察证；应当向被盘查人敬礼并告知：“我是xxx（单位）民警，现依法对你进行检查，请你配合。”盘查排除违法犯罪嫌疑后，民警应当向被盘查人敬礼，并说“谢谢你的合作”，礼貌让其离去。如果认为民警执法不规范，群众可向公安机关警务督察部门举报，也可拨打110投诉，还可以通过法律途径维护自身的权益。
+
+
+## 6.未婚同居违法吗？
+> 对其要求分割房屋的诉讼请求进行了审理，认定两人之间签订的协议没有违反社会公序良俗，属于合法有效。既然男方提出分手，依协议约定，
+
+> 法院在受理未履行登记结婚的离婚案件时，符合事实婚姻构成要件的，会认定其具有事实婚姻的效力，当事人不必再补办结婚登记手续，按离婚案件处理
+
+> 1994年2月1日以后（即民政部《婚姻登记管理条例》公布实施以后），我们国家不再承认事实婚姻，只按照“同居关系”处理。
+
+
+## 7.车借给朋友出了事故，责任由谁承担？
+> 根据我国《中华人民共和国侵权责任法》的规定，借车发生交通事故，首先由投保的保险公司予以赔偿；其次，不足部分由车辆使用人承担责任；最后，如果车主有过错，要基于过错承担责任。根据《最高人民法院关于审理道路交通事故损害赔偿案件适用法律若干问题的解释》的规定，车主有过错主要是指其明知车辆有缺陷，并且此缺陷易导致交通事故发生；知道借车人没有驾驶资格，仍然出借车辆；知道借车人缺少驾驶能力，比如饮酒、吸毒、生病等无法安全驾驶车辆等情况。
+
+
+## 9.买卖二手车未过户，发生交通事故谁承担责任？
+> 在此案张某和李某的机动车买卖法律关系中，虽然车辆未经登记，但车的所有权已经发生了转移，所有权人为张某。因此，法院判决应由张某承担赔偿责任。
+
+
+## 10.知假买假受法律保护吗？
+> 职业打假对于打击无良商家，维护消费者权益具有积极意义
+
+
+## 11.遭遇“过度医疗”怎么办？
+> 对于是否存在过度医疗的举证责任划分，法院一般遵照“谁主张谁举证”的原则。
+
+
+## 16.怎样才能确保代书遗嘱有效？
+> 法院审理查明，姜大爷具有高中文化，生前是中医，长期手写处方，不存在不识字或不能签名的问题。如果当时丧失书写能力，不得不按手印代替签名，那么需要提供现场录音、录像等其他证据予以佐证，证明遗嘱内容是姜大爷的真实意思表示。由于丁某无法提供相关证据，因此法院认定该代书遗嘱无效。
+
+
+## 22.“小三”有权分得遗产吗？
+> 不过，需要注意的是，如果是当事人与“小三”所生的子女要求继承的，应予以支持。因为根据法律规定，非婚生子女享有与婚生子女同等的权利，也即享有相同的继承权。法律不能以保护传统家庭的名义将非婚生子女排除在遗产继承之外。
+
+
+## 29.夫妻忠诚协议是否具有法律效力？
+> 法院经过审理，确认双方达成的忠诚协议合法有效，判决某甲支付某乙“违约金”30万元。
+
+
+## 36.未出生的胎儿是否享有继承权？
+> 根据法律规定，在遗产分割时，应为未出生的胎儿保留继承份额，如果没有保留的，应从继承人已经继承的遗产中扣回。
+
+
+## 42.夫妻离婚后，可以随意变更子女姓氏吗？
+> 离婚后一方如果擅自变更孩子的姓，就侵害了对方享有的子女随其姓的权利
+
+
+## 49.QQ号、微信号可以作为遗产继承吗？
+> 但是，继承人不知道被继承人的社交账户、账号及密码，而社交平台又拒不配合提供相关的信息的，继承又难以操作。
+
+
+## 50.如何约定民间借贷的利息才合法？
+> 最高法院在2015年9月1日起开始实施的《最高人民法院关于审理民间借贷案件适用法律若干问题的规定》中，对利率划出了“两线三区”。“两线”为：年利率24%（即月息2分）以下和年利率在36%（即月息3分）以上。“三区”为：年利率在24%以下的，属于“司法保护区”，法院可以认定其有效；年利率在24%～36%之间的，属于“自然债务区”，借款人有权拒绝给付，出借人不能获得胜诉权和要求法院强制执行。如果借款人自愿给付，则给付有效，事后不得要求返还；年利率在36%以上的，属于“无效区”，法院认定为无效，对于这部分利息，当事人可以在诉讼中要求返还，也有的法院会判决将已经支付的超出36%部分的利息冲抵本金。
+
+
+## 53.“订金”和“定金”，哪个可以退？
+> 订金属于预付款，而定金是签约的保证，属于“违约”定金。也就是定金不可退，订金是可以退还的。
+
+
