+++ 12.微信读书助手/导出资料/割裂世纪的战争.txt
@@ -1,0 +1,89 @@
+## 艾奇逊180度大转弯
+> 行动哲学却偏向于实用主义
+
+
+## 雷英夫预测出美军将于9月15日在仁川登陆
+> 但是，不论是斯大林，还是金日成，都没有重视这个至关重要的提醒。
+
+
+## 毛泽东作出了他平生最困难的一个决定
+> 朝鲜内战一开，解放台湾的事就只能推迟了。
+
+
+## 第一次战役（1950年10月25日-11月5日）彭德怀入朝第一天即遇险
+> 当时敌先头部队由德川经熙川窜到了我与金会谈的大洞东北方向的桧木洞，已绕到我们住的大洞后边去了。我志愿军刚过江不远，即与该敌遭遇，我与金日成幸免被俘
+
+> 而彭氏却在中国历史上留下了一言难尽的悲剧式的光荣
+
+
+## 威洛比少将断定“山那边的敌人不会介入战争”
+> 中国人会不会出兵？”麦克阿瑟非常有把握地回答，绝对不会。显然，此时的他并非像他后来在回忆录中所说，对中国和俄国的介入都看得一目了然。
+
+> 迷信产生的盲从取代了威洛比最初的清醒
+
+
+## 一枚凝固汽油弹击中了毛岸英
+> 彭德怀一声令下，几十万志愿军与美李军脱离接触。在回撤路上，中国军队有意丢下了许多被装，一副“仓皇北窜”的景象。同时，还释放了100多名俘虏，他们被告知志愿军供应困难，要回国了。接着大军就像蒸发了一样，消失在北朝鲜白雪皑皑的山里。显然，这是在“有意示弱、诱敌深入”，彭德怀采用中国最古老的兵法，来对付最现代化的美国军队。
+
+> 最初，在敌机轰炸时彭德怀也毫不躲避，他对身边人员说，“谁怕死，就去躲飞机”。也许他还是依照对国民党军队的看法，去看待美国空军的轰炸。惨痛的教训很快就来了。
+
+> 11月25日拂晓，在听到防空警报后，洪学智以休息下棋的名义，硬拉着彭德怀到山洞躲避。几分钟后，一组B-26轰炸机轰炸了设在大榆洞的志愿军总部。有几颗炸弹命中了彭德怀住的独立屋。离独立屋不远处，在办公室的毛岸英，不幸未能躲过凝固汽油弹的袭击。
+
+> 50年前中美两国在朝鲜作战并不是被狭隘的民族利益驱动，而是由于冷战时期的意识形态对抗所致。这真是大国的不幸。
+
+
+## 第二次战役之西线（1950年11月25日）梁兴初咬着牙说这次要打他个狠的
+> 因此在前进中一听到枪炮声就停止不前，只是喜欢昼间行动，甚至不习惯于夜战和近战。打了败仗不能保持秩序……如果断绝了补给立即就会丧失战斗意志。
+
+
+## “朝另外方向进攻”的史密斯将军
+> 当发现在水门桥附近已有志愿军据守时，负责占领高地的美军第7团索耶少校大吃一惊。他意识到这将是一场死战。可是索耶少校组织部下对志愿军阵地进行攻击时，居然未受到丝毫的还击。占领高地的美军看到了令人惊讶的一幕：志愿军的士兵已全部冻僵在堑壕之中！
+
+
+## 第三次战役（1950年12月31日-1951年1月7日）史蒂柯夫大使要彭德怀“把敌人赶下海去”
+> 史蒂柯夫大使曾经担任过苏军驻朝鲜集团军的司令，也许是将军出身，所以此公格外喜欢就军事问题发表看法。史蒂柯夫对彭德怀在二次战役的指挥感到不满，他认为美军正在迅速向三八线撤退，志愿军应该乘胜追击，像苏联红军打关东军一样，十几天解决战斗。
+
+> 彭德怀的各项分析都极其清醒、极其准确，绝非史蒂柯夫之辈所能企及。
+
+
+## 周恩来的“官方明信片”
+> 在1951年1月，中华人民共和国被排除在联合国之外，那种为国家生存而战的孤独感，是已经“加入国际组织并在其中担负重要职务”的今天的中国人难以理解的。就像在优雅的亚布力高山滑雪场穿着鸭绒衣的游客无法体会穿单裤和胶鞋在长津湖冰面上潜伏的志愿军战士的感受一样。
+
+
+## 第四次战役（1951年1月25日-4月21日）李奇微发现了“阿喀琉斯之踵”
+> 李奇微通过思考，终于得出了一个了不起的对志愿军来说几乎致命的结论
+
+> 李奇微无疑是一个出色的战术家。作为战场指挥官能成为战术家是一种很高的境界。直到今天，他对当面敌人研究的方式，对每个战场指挥官仍有教益。
+
+
+## 杨迪在横城捡了一次“洋捞”
+> 因为彭德怀此次北京之行，与毛泽东、周恩来等中央政治局常委谈了“我们的朋友和同志对不乘胜追击”的看法，谈了志愿军作战的困难，谈了李奇微战术的变化，谈了要组织第二番部队入朝作战，谈到应暂时避免战役性出击等广泛话题，使毛泽东等中国领导人对朝鲜战争的艰难程度有了确切认识，对盟国间在战场上的复杂关系有了更深的认识，从而确立了新的战略方针。
+
+
+## “烂醉如泥的夜晚”杜鲁门下达了撤职令
+> 在解除麦克阿瑟的职务这件事上杜鲁门做得太中规中矩，不但没有丝毫的错处而且还显得是麦克阿瑟一而再、再而三地犯错，在美国历史上还没有一位司令官能像他那样多次抗命，让人感到解除麦克阿瑟的职务实在是不得已而为之
+
+
+## 天王老子也不怕的王近山疯不起来了
+> 三兵团六十军一八〇师在后撤中被美军“特遣队”隔离，师领导误以为被包围，故决定全师分散突围，除师长郑其贵率一个建制营和几小部分部队突围成功外，政治部主任吴成德以及7000余人战死或被俘，写下了志愿军战史上最难以启齿的一笔。
+
+
+# 第三部 无法打破的僵局
+> 1959年，邓华因彭德怀事件牵连被撤职。“文革”后获平反，不久去世。
+邓华是继彭德怀之后，中国人民志愿军的第二任司令。今天如果我们用一句话来概括邓华在朝鲜战争中的作用，那就是他的作用要比现在人们知道的都要大。
+
+
+## 克拉克将军“随时准备着”
+> 坚决反对停战谈判的李承晚，于4月9日向美国总统递交了一封抗议信，坚决反对恢复谈判。
+
+
+## 给李承晚最后一击的杨勇
+> 李承晚为破坏停战谈判，6月17日夜，以“就地释放”为名，强行扣留了2.7万名朝鲜人民军被俘人员，并把他们中的许多人强制编到南朝鲜的军队中去。紧接着，李承晚宣布，停战协定一旦签订，“要把我的军队从联合国管辖下撤出来”，“要打到鸭绿江”。
+
+
+# 结语 越过战争的肩头向更远方眺望
+> 从最初的错误，到最后的错误；从构成原因的错误，到导致结果的错误；从大错误，到小错误——一场整整打了三年零两个月的战争，简直就是一个错误的集合体，
+
+> 假如当年雅尔塔的“巨头们”不轻率作出把朝鲜一分两半的决定；
+
+
