+++ 12.微信读书助手/导出资料/半小时漫画世界史.txt
@@ -1,0 +1,105 @@
+# 序
+> 公共史学的兴起
+
+> 历史是过去的政治，政治是当代的历史。
+
+> 这是一本极简版的世界史，半小时即可读完，但有限的文字和篇幅却浓缩了世界史最重要的内容，包括基本历史线索、重大事件和重要人物；该书图文并茂，文字生动，加之大量使用时尚的网络语言，特别适合年轻人阅读。
+
+
+# 一、简到崩溃的欧洲史（一）：希腊 & 罗马
+> 简到崩溃的欧洲史
+
+> 希腊文明1.0时代：克里特文明。
+
+> 隔壁亚洲冒出来个巨大的帝国叫波斯。老板很喜欢吃又很抠门，喜欢吃希腊爱琴海的鱼，又嫌进口贵，想来想去
+
+> 大敌当前，雅典和斯巴达秒变好队友：
+
+> 伯罗奔尼撒战争。
+
+> 马其顿
+
+> 执政官+元老院”这个男子组合共同执政，
+
+> 这个皇帝就是——君士坦丁大帝。
+
+> 东罗马继续活了一千年，而西罗马不到一百年就让外邦人灭了。
+
+
+# 二、简到崩溃的欧洲史（二）：黑暗的中世纪
+> 扑克牌里的这张红桃K，就是查理曼大帝。
+
+
+## 一、神圣罗马帝国
+> 第二帝国是普鲁士时代。语文课文《最后一课》，说的就是法国被普鲁士占领的情形。
+
+
+## 二、英法战争
+> 英格兰的国王，在法国也有地盘，比如著名的诺曼底。
+
+
+## 三、十字军东征
+> 后来他们还抢了一块很重要的地段：耶路撒冷。
+
+
+## 四、文艺复兴
+> 中世纪景象四件套：文化全靠背经，艺术只画上帝一家亲，科技全赖村头铁匠铺，外交就是不断打中东。
+
+
+# 三、简到崩溃的欧洲史（三）：下海的欧洲
+> 奥斯曼帝国
+
+
+## 一、大航海时代开始啦
+> 从此它的地盘遍布全世界，24小时都能见到太阳，号称日不落帝国。
+
+
+## 二、同去革命不？同去同去
+> 路易十四，跟咱们康熙帝一个时代，一样威武，而且是个文艺青年，最喜欢在舞台上扮演太阳神，所以有人称他太阳王。
+
+
+# 四、屏幕中的历史——《斯巴达300勇士》（上）
+> 波斯大家都知道，就是现在伊朗的前身
+
+
+# 五、屏幕中的历史——《斯巴达300勇士》（下）
+> 美男子一怒之下，一把火烧了雅典。
+
+
+# 六、十字军东征 之 三大骑士团
+> 后果是圣殿骑士团全被搞死，那天是10月13日星期五，也变成了“黑色星期五”的说法之一。
+
+
+# 八、美国往事（一）：独立战争
+> 纽约的自由女神像，就是法国人在美国独立100周年的时候送的，象征自由和反抗暴政。
+
+
+# 九、美国往事（二）：南北战争
+> 最早的美国，只有十三个州。
+
+
+## 二、南北战争
+> 因为像美国这种动不动搞投票的国家，阵营越大，就越能控制国家政策。
+
+> 历史上大多数的“正义”，通常都并存着非常直接的利益冲突，所以大家在理解历史、理解世界的过程中，要正确认识人性的作用，不要贬低也不要拔高。当然，我们没必要因此对世界失望，即便利益是每场战争的重大影响因素，人性的真善美总能在大方向上掌控全局。
+
+
+# 十、美国往事（三）：上帝保佑美国
+> 日本人的谜之顽固，使自己变成了世界上唯一一个被原子弹炸过的国家
+
+
+# 十一、一口气读完日本史
+> 东京不要说去东京，要说去江户；去京都不要说去京都，要说去平安。
+
+> 先别急着听大帅哥织田信长和猴子秀吉的战国故事，他们活在16世纪，同时期是我们的大明朝。
+
+> 大和民族
+
+> 武士，本来是一个下层群体，但天皇自己没本事，得求他们来保护自己的时候，武士集团开始厉害起来了。
+
+> 这就是日本激动人心的战国时代。
+这个时代出现了著名的三连发：织田信长、丰臣秀吉、德川家康。他们依次坐庄，最后统一了日本，结束了乱世。
+
+> 整个幕府时代，天皇一直没啥权力，这就是为什么下面大家打得再嗨，也没人上去动他——动他也没啥用啊。所以天皇这个工种，是祖传的，一家人从头到尾没断过，所谓“万世一系”。
+
+
