+++ 12.微信读书助手/导出资料/半小时漫画中国史2.txt
@@ -1,0 +1,79 @@
+# 一、大汉王朝之西汉流水账
+> 于是以后只要和匈奴出点什么问题，就安排和亲。
+
+> 几年面子是挣足了，家也差点败光了，可是汉武帝命好，爷爷爸爸辈给他攒钱，还有小字辈替他还债。
+这就是第二个黄金搭档：汉昭帝+汉宣帝
+￼
+
+> “外戚政治”。
+
+
+# 二、大汉王朝之王莽篡汉
+> 治国有时候和超级大脑是不兼容的，当思想超前时代太多，就是无土之木、无源之水
+
+> 在“绿林军”中，有个叫刘秀的，祖上是汉武帝的兄弟，也算皇亲国戚，可是到他这辈已经混成了平民。
+
+
+# 三、大汉王朝之东汉故事
+> 这事没完没了地持续到东汉末年，叫“党锢之争”。
+￼
+
+> 中国历史上规模最大（波及全国）的一次农民起义——黄巾起义。
+
+
+# 四、大汉王朝之西域篇
+> —乌孙
+
+> 匈奴人没文化，不会写日记，所以没留下记录，所以没有自己的历史，被赶跑以后就消失了。
+
+
+# 五、三国之前戏要做足
+> 这就叫挟天子以令诸侯。其实，令诸侯是假，令不动就代表皇帝消灭你，这才是目的。就这一步，曹操的政治智慧就甩了大家二十八个袁绍的距离。
+
+
+# 六、三国之三大战役（上）
+> 关东联军算是黄了，但袁绍自己抢地盘还是可以的，散伙后在河北发家，最后击败了公孙瓒，变成了北方一哥。
+
+> 一场东风刮过，孙权放一把大火把曹操连起来的战船烧了个透，号称几十万大军，就搁这儿打了水漂
+
+
+# 七、三国之三大战役（下）
+> 赤壁之战以后，曹操灰溜溜地回到老家，天空失去颜色，悲伤逆流成河，再也不敢随便南下找碴儿了。
+
+> 一代传奇关二爷，就这么死了，这就是“大意失荆州”和“败走麦城”。
+
+> 老刘家这时候顶梁柱基本全倒了，就剩刘备的傻儿子刘禅，没啥好说的，直接灭掉。
+
+> 还剩下江东吴国，孙权已经死了，现在的老大是他孙子孙皓，残忍无良，也是一个活不过两集的角色。
+
+
+# 八、三国之哪儿来的东风
+> 一种意犹未尽的感觉无声又无息，出没在心底对不对？
+
+> 赤壁刮的东风是咋回事？
+答案就是：
+我也不知道！
+
+
+# 九、两晋之西晋有点乱
+> 这就是司马懿，一个听音乐会从来不敢坐前排的高智商男。
+
+> 不过他不是主角，只是道具。
+这个司马衷的智商跟别人不一样，特别低。
+
+> 用一杯满是金屑的酒灌死了。
+贾南风死后
+
+
+# 十、两晋之五胡入华
+> 苻坚。
+他消灭了大致并存的燕、凉（西北）等国家，第一次统一了北方，然后跟东晋进入著名的南北部总决赛——淝水之战。
+
+
+# 十一、两晋之东晋真纠结
+> 玄立即把白痴皇帝司马德宗赶下来，自己当皇帝，结果前脚上位，后脚就被人赶了下来。
+
+> 这人不但搞定了桓玄，手一滑连整个东晋也灭掉了，他就是晋帝国的终结者——刘裕。
+从此中国的南方开启一个新的朝代，史称南朝。
+
+
