+++ 12.微信读书助手/导出资料/半小时漫画中国史3.txt
@@ -1,0 +1,27 @@
+# 二、隋朝二人传
+> 历史上把杨坚开创的巅峰时期称作开皇之治。
+你也可以喊它杨巅峰。
+眼看着剧情一路走高，别急，杨坚在巅峰的时候得了一种病，叫妻管严。
+
+> 只要他一出门，国内就造反，因为大家实在受不了这么打仗了啊！有一次他差点就攻破了高句丽的城墙，结果国内又反了，只好吭哧吭哧跑回家，两边亏。
+
+> 就在这个时候，一个小弟瞅准时机反水，把隋炀帝缢死了。他就是隋炀帝的小伙伴宇文化及。
+
+
+# 三、隋唐英雄传
+> 虎牢关大战
+
+> 本来老李家这次只想收拾王世充一个，窦建德以后再说，谁知道窦建德积极配合，一炮两响，天下忽然就平定了！
+
+
+# 八、大唐就这么挂了
+> 比如一个姓牛的和一个姓李的，拉帮结派搞内斗。
+这就是非常闹心的牛李党争。
+
+> 黄巢起义
+
+
+# 九、大唐盛世番外篇——大唐长安城，怎么这么大！
+> 但是有繁荣就有衰落，长安城的繁荣和大唐王朝的鼎盛是分不开的。等到安史之乱之后，大唐一天不如一天，长安城也就慢慢衰落下来，最后在连年兵乱中被逐渐废弃。
+
+
