+++ 12.微信读书助手/导出资料/半小时漫画中国史4.txt
@@ -1,0 +1,47 @@
+# 一、五代十国广场舞
+> 后周作为五代的压轴，也是五代中最强的朝代。郭威自己就是个很厉害的皇帝，而他的养子柴荣，也就是周世宗，更是五代十国里最优秀的皇帝，没有之一。
+
+
+# 三、大宋风云（二）：文艺青年宋徽宗上台
+> 一个坐过站的大妈是很可怕的，因为她会带着所有人跑偏
+
+
+# 四、大宋风云（三）：北宋被逼成了南宋
+> 因为虽然皇帝在杭州，但没有放弃北方，所有人心目中真正的首都，还是开封。
+
+> 撼山易，撼岳家军难!
+
+
+# 五、大宋风云（四）：彪悍的蒙古人来了!
+> 男人温柔像条狗，不是心机就是丑。
+
+> 堂堂金国皇帝，听了一句宋朝网红的歌词，然后败给了一个参谋，最后就这么死在了长江边。
+
+> 铁木真统一蒙古之后，立刻找金国显摆了一下实力，谁知昔日的东北大哥金国，一上来就被打蒙了。
+
+> 不过蒙古人没有立刻把金国灭掉，打到一半跑到中亚，抽空揍了一把西辽和花剌（là）子模，这部分我们将在之后的蒙古帝国篇详述。
+
+> 不用说，以宋朝的尿性，果断放弃金国。公元1234年，南宋跟蒙古一起，终于把金国给灭了。
+
+> 而宋朝亲手送走了辽国，又送走了金国，现在该面对蒙古帝国了。
+
+> 当时攻打南宋的主力将领就是忽必烈，他为了赶回蒙古抢大汗的宝座，只能答应贾似道议和，然后撤退了。
+
+
+# 十、元朝番外篇——元朝为啥这么短命？
+> 无数的例子告诉我们，一个彪悍的汉子突然变得弱不禁风——
+如果不是去了泰国，就一定是自身的命脉出了毛病!
+
+> 元朝的纸币还曾一度在东南亚流通过，有点像亚元的感觉。
+
+> 元朝的短命，看似是因为政治问题，但其实是因为经济问题。
+
+
+# 十一、元朝番外篇——风靡娱乐圈的元曲
+> 元散曲非常口语化，根本不用翻译，通俗得都快掉渣了。
+
+> 科举制一直是中国文人的命根子，要想不被同龄人抛弃，就得玩命学习，考科举。
+
+> 但是到了元朝，有80年的时间都没有举行过科举考试。
+
+
