+++ 12.微信读书助手/导出资料/半小时漫画中国史番外篇：中国传统节日（独家定制版）.txt
@@ -1,0 +1,84 @@
+# 一、你可能不知道，春节其实也就100多年的历史
+> 大家是不是也一直很费解——
+春节每年都过，但为啥具体的过年时间却总是不确定呢？
+
+> 此时出现了个高科技——圭表测影。
+
+> 传统意义上的春节，是指腊月初八的腊祭或腊月二十三的祭灶，到正月十五，其中除夕和大年初一，是整个春节的高潮。
+
+> 人们一开始过年没后来这么多花头，什么贴春联、放鞭炮都是不存在的。
+那时技术不发达，天气还特别坏，种地非常辛苦。于是，到了年终，大家就开始求助各路神仙，希望来年有个好年景。
+
+> 在门上挂上桃符避避邪，据说这种画着门神的小木片，慢慢就演变成了现在的春联。
+
+> 春节的时间则被定为腊月二十三到正月十五。
+
+
+# 二、元宵节：吃什么汤圆？!在古代，那可是情人节
+> 从此之后，每年的正月十五，家家都会张灯结彩，来纪念诸吕之乱的平定。
+
+> 于是他就觉得，自己应该表示一下，比如写封感谢信、送个锦旗啥的。
+
+
+# 三、清明节：古人过清明，那是相当欢乐
+> 在很久之前，清明节是个愉快的日子。
+
+> 上巳节一般在农历三月三，相传是为纪念黄帝诞辰，是中国最古老的节日之一。
+
+> 二、寒食节
+寒食节一听名字就比较衰了，一般在冬至后105天
+
+> 然唐朝的三观各种前卫，但这么干还是有点太刺激了，类似全民直播坟头蹦迪
+
+
+# 四、端午节：屈原他自己都过端午节
+> 端午节是咱们重要的传统节日，在2009年，还被联合国列为非物质文化遗产，是中国第一个入选世界非遗的节日。
+
+> 甜粽子和咸粽子你选哪一个？
+
+
+# 五、七夕节：除了谈恋爱，你还应该知道啥？
+> 七夕节这天，有很多人求交往，天地间爱意弥漫，因为牛郎织女的狗血爱情故事流传了这么多年，实在深入人心。
+
+> 最初的七夕压根就跟爱情没关系。
+
+> 而在我们中国，古人就将黄道和天赤道附近的星空，分成二十八个地盘，作为日月星辰的“宿舍”，就叫二十八星宿(xiù)。
+
+> 而这种祈祷心灵手巧的活动就叫乞巧，所以七夕又叫乞巧节。
+
+> 再加上媒体商家等社会各界的推动，如今的七夕已经是名正言顺的情人节了。
+
+
+# 六、中元节：鬼节这天能不能上街？
+> 鬼是怎么来的？
+“鬼”这个词，出场自带封建迷信色彩，但如果我们把它当成一种文化来理解——
+
+> 这种看不见摸不着的灵魂，就是最初鬼的概念。
+
+> 鬼在古人眼里那可是集可爱与智慧于一身的存在，还能帮助人实现心愿、保佑平安
+
+> 东汉有个大V叫张道陵，创了个五斗米道，后来发展成道教。
+
+
+# 七、中秋节：你中秋赏月赏了个啥？
+> 拜月当然得挑月亮又大又圆的时候，也就是八月十五
+
+> 小知识：据说在清朝，因为“饼”与“病”同音，所以人们为了避讳，就把月饼叫作月华。
+
+> 所以，月亮都长成这样了，你还赏啥啊!!
+
+> 月亮的变化，也就是我们所说的月相。
+
+> 虽然月亮素颜挺丑，但是能跟家人一块儿欣赏，管他呢!
+
+
+# 九、腊八节：别光顾着喝粥，得知道为啥喝!
+> 因为娱乐活动少，所以只有官方才重视腊日节。到了唐末，连官方也不注重腊日节了
+
+> 现在知道这一碗腊八粥里都包含着啥了吧？
+
+
+# 十、二十四节气——古代劳动人民的大智慧
+> 这个把昼夜平分的神仙，就叫春分。
+
+
