+++ 12.微信读书助手/导出资料/半小时漫画理财课.txt
@@ -1,0 +1,36 @@
+## 第2课 国债
+> 而低风险投资，如果能长期坚持，通过不断的利滚利，
+是可以给你带来高收益的。
+比如，国债。
+
+
+## 第3课 P2P网贷
+> 但是对于大多数投资者来说，P2P网贷仍是性价比最高的中短期固收产品。
+
+
+## 第5课 基金定投
+> 在现实情况中，人们往往喜欢看到A这样的行情，越涨越买；而对C非常抗拒，甚至恐慌，越跌越不敢买，却不知失去了获得正收益的宝贵时机。理性投资从来都是反人性的，因此赚钱的永远是少数人。切记！切记！
+
+> A类：前端收费，即申购的时候收取，是基金最常见的收费模式。
+B类：后端收费，即赎回时收取，一般来说B类比A类更划算。
+C类：申购和赎回时均不收取费用，仅在持有期按日计提较低比例的销售服务费，适合1年内的短期投资。
+
+
+# 下篇 专业课
+> “即使我能洞悉地心引力，
+也无法计算出人类的贪婪和恐惧。”
+
+> 我们把在上海、深圳交易所上市的股票，叫“A股”；把在香港联交所上市的股票，叫“H股”。
+
+
+## 第7课 债券
+> “穷人炒股越炒越穷，富人买债越买越富。”
+
+> “富人都买债，穷人才炒股！”
+
+
+## 第8课 黄金
+> 手里钱不多的朋友，要谨慎考虑黄金投资。
+支持是因为黄金特有的避险功能，时局越是动荡，黄金越是受欢迎。在其他资产下跌的时候，唯独黄金价格能上涨。黄金的这个作用是很难替代的。
+
+
