+++ 12.微信读书助手/导出资料/半小时漫画经济学2：金融危机篇（独家定制版）.txt
@@ -1,0 +1,157 @@
+# 一、开篇：我们为什么要了解金融危机
+> 就是1929年金融股市大崩盘引发的美国大萧条
+
+> 大萧条的破坏力如此之大，以至于美国人民连饭都吃不上。然而万万没想到，一个天大的机会随后砸来，让美国人成功咸鱼翻身。
+
+> 当然，二战之所以能打起来，部分原因是美国的国内危机，演变成了全球危机。
+
+> 大批量的空房子又刺激了房价进一步下跌，不想还房贷的人更多了。
+和大萧条一样，美国又一次将危机传染给了全世界。
+
+> 2012年，天猫在“双十一”的总成交额是191亿元；而到了2018年，这个数字已经达到了2135亿元。
+
+> 一战、大萧条和二战的发生都是因为没发明网购；次贷危机的真相不是房价暴跌；金融危机不是绝对的坏事，也可能是社会的崩溃疗法……
+
+
+# 二、金融危机就要来了：明斯基时刻
+> 人们把这种周期性经济衰退的现象叫：经济危机。
+
+> 他叫海曼·明斯基，提出了“金融不稳定假说”
+
+> 先还着利息，以后再还本金，养鸡场押给你。
+
+> 多借钱等于多赚钱
+
+> 借的这部分钱，就叫杠杆。借钱用来扩充资产，就是加杠杆。杠杆让你赚得多，但赔起来也多。
+杠杆越来越高，最后还不起了，可能就会引发危机。
+
+
+# 三、一朵花酿成的惨案：郁金香泡沫（上）
+> —郁金香泡沫，堪称金融危机鼻祖。
+
+> 荷兰的历史就是一部以经济为中心的历史。
+
+> 东印度公司权力相当大，能印钱有军队，能跟国家签合约，但它最大的使命就是赚钱给股东分红。
+
+
+# 四、一朵花酿成的惨案：郁金香泡沫（下）
+> 荷兰崛起之后的这种疯狂状态，大概就是郁金香泡沫的背景。
+
+> 首先郁金香很美，大家都喜欢；
+其次，它很难种。
+
+> 一种东西，如果能同时兼具美和少，那它一定可以成为奢侈品！
+
+> 于是市场上的需求，从好想要一朵郁金香，变成了好想要一朵与众不同的郁金香。
+
+> 原来郁金香的价格天天在涨，今天一块钱，明天十块钱。这就吸引了一批人，低价买，高价卖，来赚差价。
+这批人就是投机者。
+
+> 合同到期了，就能拿到郁金香，反正不会亏，大家都乐意。这种玩法就是期货的雏形。
+
+> 一朵花值这个价吗？对社会有贡献吗？但是走上人生“癫疯”的荷兰人并不这么想，他们砸锅卖铁，放下工作，兢兢业业去炒郁金香。
+
+> “泡沫”这个词儿很形象，它有一个明显的特点：
+￼
+三、早该知道泡沫，一触就破
+
+> 但是泡沫这种东西，一扯到大妈，有没有感觉随时都会破掉？
+
+> 荷兰人全国总动员炒郁金香，船不造了，生意黄了，钱投进去连个响都没有，经济急剧下滑。
+
+> 17世纪30年代后期，荷兰的经济是真不行了，一下子被英国超了车，海上扛把子地位不保。
+
+> 起码你再看到有些东西很贵，但明显不值那个价，人们却一直往里砸钱投资时，就应该知道那很可能就是金融泡沫，要保持冷静。价格虚高的东西别碰，要投就投有价值的，这叫理性。
+
+
+# 五、连牛顿也算不出的疯狂：南海泡沫
+> 宁可相信世上有鬼，也不能信人类的嘴。
+
+> 当时的欧洲人没啥文化，
+所以就把南美洲叫南海。
+
+> 事情是这样的：西班牙国王有个姐姐，嫁到了法国，因此西班牙国王就是法国国王的小舅子，法国国王就是西班牙国王的亲姐夫。
+
+> 奥地利国王和西班牙国王是同宗，祖上一个妈生的那种，于是奥地利国王跳出来说，你姐嫁出去就是外人，这个沙发应该归我儿子才对！
+
+> 英国政府琢磨着，老百姓有钱啊，于是想了个金融办法：
+发国债！
+
+> 借钱一时爽，还钱火葬场
+
+> 亚洲有香料，美洲有黄金啊
+
+> 此后，南美洲归了西班牙人，谁想进来赚钱，都得他们点头才行。
+
+> 西班牙规定，英国每年只能派三只船过去贸易。
+
+> 南海贸易特许权本来就是个金字招牌，再加上政府的担保，就能让不明真相的老百姓信得不要不要的。
+
+> 据说英国政府本来是反对南海计划的，但土豪们贿赂了官员，使这个计划得以通过。
+
+> 于是大家一窝蜂开公司，出现了很多奇葩公司，疯狂发股票忽悠人买。
+
+> 泡沫法案
+
+> 就连大科学家牛顿也赔得老惨了！为此牛大爷还发微博叹息道：
+
+
+# 六、股票连涨13个月会怎么样：密西西比泡沫
+> 市场没钱，消费力低下，物价下跌，这种现象就叫通货紧缩。
+
+> 没错，他死了……
+
+> 这些操作说白了，就是以国家为担保，让大家相信纸币的价值。
+老百姓一琢磨，国家都出面了，只要纸币能买东西就行，于是大家开心买买买，纸币开始流通。通货紧缩的难题就这样解决了！
+
+> 政府出面配合说，这公司有特权，只有它能去美洲。然后约翰大叔开始给老百姓洗脑
+
+> 于是密西西比公司的股票，连续上涨了13个月，天天都是涨停板。
+
+> 钱是多了，但东西就那么点，因此物价要涨，钱就不值钱了。
+￼
+这就叫通货膨胀。
+
+> 法国经济繁荣靠股市，但股市其实很虚，因此所谓的经济繁荣就是个泡沫啊！
+
+> 约翰·劳的墓志铭中写着这样一句话
+
+> 经济学得好可能证明不了啥，可一旦学不好，后果很严重啊！
+
+
+# 七、房价是怎么被炒起来的：1837年美国大恐慌
+> 可能是州银行饿坏了，本来应该有多少黄金，就印多少纸币，结果爽过了头，一不小心纸币印多了。
+
+> 美国房地产根本不值这个价，全是炒出来的，供过于求，市场上没那么多需求，都是虚的。
+
+> 这才发现，很多州银行都滥印了纸币，根本没那么多黄金。
+全国人民心态瞬间崩成渣渣，银行也被挤兑到破产了。
+
+> 1848年，恐慌了11年的美国在加州的一个地方，发现了大金矿。本来缺黄金，一下子却变成了矿主，美国突然不慌了，这就是历史上著名的淘金热。
+
+
+# 八、美联储的诞生：1907年美国银行危机
+> 还记得我们曾经说过，一个东西不值它的价格，高价都是被炒上去的，这是啥？
+
+> 越来越多的人有了这种心态，一窝蜂跑去银行取钱，这叫挤兑。
+
+> 当然摩根也不是单打独斗，背后还有政府的支持。
+银行有了足够的钱，就不怕被挤兑破产了；股市上有了“接盘侠”，股价也开始逐渐升温。很快，美国银行危机就解除了。
+
+> 不管你愿不愿意承认，生意场没那么多无私奉献救人于水火的超级英雄。一个精明的资本家，愿意拿出那么多钱，投在已经非常危险的市场，如果不是在做慈善，那肯定还是因为利益。
+
+
+# 九、实现美国梦的三大法宝：泰罗制、福特制和猪
+> 最后美国还参战成了战胜国，又捞了一笔赔款。
+
+> 弗雷德里克·泰罗被称为科学管理之父，他的泰罗制内容很多，咱只挑重点讲。
+
+
+# 十、1929—1933年美国大萧条
+> 垄断的意思大家都知道吧：一家公司赶走竞争者，占领市场一家独大。
+
+
+# 十一、拯救美国的两位“救市主”：胡佛和罗斯福
+> 下滑期，再厉害也没用；上升期，躺赢不是梦。
+
+
