+++ 12.微信读书助手/导出资料/半小时漫画经济学3：金融危机完结篇（独家首发）.txt
@@ -1,0 +1,98 @@
+# 一、中东土豪复仇记：石油危机(上)
+> 在亚非欧三大洲之间，卡着一块神奇的土地，它就是中东。
+
+> 石油是怎么来的，其实有很多种说法，这里说的生物沉积变成石油，是其中的一种。
+
+> 煤油不是从煤里来的，而是从石油里面提炼出来的
+
+> 据说是因为他后来不去挖石油，改行炒股了。
+
+
+# 二、中东土豪复仇记：石油危机(下)
+> 但凡上过初中的都知道，班里谁牛气，并不是看个头大小，而是看他哥是谁!
+
+> 各国都得用美元买石油，囤的美元一多，最后还得去美国消费，这就叫石油美元回流。
+
+
+# 三、不过是签个字，经济就停滞了十年：日本与《广场协议》(上)
+> 美国之所以帮助日本，主要原因是：
+日本商品物美价廉，买日货既能满足消费需求，又能帮日本，何乐而不为？
+
+
+# 四、不过是签个字，经济就停滞了十年：日本与《广场协议》(下)
+> 因为这次的利率降得太低了，低到不向银行借点儿钱，都觉得跟丢钱了一样。
+
+> 日本还在美国狂购公司、地皮等，甚至买下了纽约的帝国大厦。
+
+> 欲让其灭亡，先让其肥胖膨胀。
+
+> 结果科威特被干趴下了。这事激怒了联合国，于是对伊拉克进行了一系列制裁，其中就包括禁止石油出口。
+
+> 石油再次减少，就导致了第三次石油危机。
+
+> 三重野康，日本央行行长，
+被称为戳破泡沫的人，
+长年占据“日本最讨厌的人”
+TOP榜前列。
+
+> 人们一看到跌价就会疯狂大甩卖，所以最后股价房价一落千丈。
+
+> 而泡沫破灭的时间，刚好也是平成时代，于是这代人成了大家口中的“平成废物”。
+
+
+# 五、一场由程序员“背锅”的股灾：1987年的黑色星期一
+> 纽约证券交易所，简称纽交所
+
+> 最夸张的说法认为，这次的股灾要归咎于程式交易和程序员。
+
+
+# 六、三步搞翻亚洲经济：索罗斯与亚洲金融风暴(上)
+> “亚洲四小龙”分别是中国香港、中国台湾、新加坡、韩国；
+“亚洲四小虎”分别是泰国、印尼、马来西亚、菲律宾。
+
+> 亚洲片区能突然有如此高速的发展，最本质原因是投资和劳动力的增加。
+
+
+# 七、三步搞翻亚洲经济：索罗斯与亚洲金融风暴(下)
+> 金融巨鳄”，趁你虚要你命的杰出代表，K.O英国央行的纪录保持者
+
+> 很明显，做空就靠赚差价，所以降价是关键。
+
+> 借到泰铢之后，他马上把它们全扔到市场上进行抛售。
+
+> 让市场上的泰铢变多，这是索罗斯让泰铢贬值的关键一步!
+
+> 可惜外汇数量有限，实在不够和索罗斯对抗。
+
+> 更雪上加霜的是，泰国国内有一撮人贪图利益，选择主动站在索罗斯这边。
+
+> 原本对自己的技术迷之自信的索罗斯，这次却差点儿连牌桌都没下来。
+
+> 联系汇率制，
+
+> 在当年，中国中央政府和香港特区政府的外汇储量加起来排世界第一。
+
+> 有了中国中央政府的支持，香港特区政府的信心大增，开始有计划地抬高股价，稳定港币，情况因此发生了逆转。
+
+> 但索罗斯却不这么认为，相反，他觉得自己无比正确，还有这么一套说辞：
+￼
+在金融上，说不上有没有道德，都只是一种操作。
+——乔治·索罗斯
+
+
+# 八、被买房压力逼出来的次贷危机(上)
+> 像这种要啥没啥、人品不佳、借了钱不一定能还上的人借的贷款，就叫次贷。
+
+
+# 九、被买房压力逼出来的次贷危机(下)
+> 这个“息”不是人们向银行借钱时产生的利息，而是银行向美联储借钱的时候产生的利息。
+
+> 还不起的这种情况就叫断供。
+
+> 次贷危机
+
+
+# 十、结论：金融危机生存指南
+> 在如今全球一体化的大背景之下，任何国家发生了危机，都会产生不可估量的影响。
+
+
