+++ 12.微信读书助手/导出资料/半小时漫画经济学（第1册生活常识篇）.txt
@@ -1,0 +1,98 @@
+# 一、开篇：我们为什么要读经济学
+> 1、宏观方面
+不管你愿不愿意承认，历史上没有一场仗是因为爱和正义打起来的，所有的流血背后，深层次的原因都是经济。
+
+> 其实很大程度是经济原因。官府横征暴敛，征的是税，敛的是老百姓财产。
+想象一下，你家的钱全被皇帝收走去泡妹子了，你会咋做？
+
+> 比如牛顿：
+你只知道他是个物理学家，却不知道他还是英国铸币厂高管，提出了金本位。
+
+> 看懂了亚当·斯密的隐喻——看不见的手。
+
+> 首先，经济是什么？
+简单地讲，就是资源最优化配置。
+
+> 所以姑娘们，看明白了吧，以后男朋友不给你清空购物车，就把这本书甩给他！
+
+> 严复曾将其翻译为“生计”，后来因为经济这个词符合时代潮流，就改为了经济。
+
+> 但能明白生活的逻辑，运用这些知识，去解决问题！
+
+
+# 三、那些年缴过的“五险一金”到底是什么
+> 说好税前一万五，发完工资只够吃土，
+
+
+# 五、医保到底保了个啥
+> 有医保是好事，可有时去报销才发现，这也不能报，那也不能报……
+
+> 非治疗类或第三方造成的伤害，一般医保不能报销，比如整形、保健、车祸、打架等。
+
+
+# 六、一口气搞懂到底要缴哪些税
+> 有人觉得：工资低，连税都没缴过，还算纳税人吗？
+
+> 话说哪些个人收入需要缴税呢？
+
+> 叫了这么多年的“起征点”，其实大家都叫错啦！
+正确叫法应该是免征额，
+
+> 我们发现，挣钱得缴税，花钱也得缴税，那只挣不花，会怎么样？
+
+
+# 七、买房前必须知道的二三事
+> 咱们知道，在旧社会，大多数土地集中在少数人手里，比如地主
+
+> 农民把土地交给国家，建起了合作社和人民公社，这就是集体生产。
+
+
+# 八、人口与经济：生娃原来是个经济问题
+> 原始社会时期，大家的精神世界很贫瘠，一到晚上没什么娱乐，就只能活动活动。
+
+> 人多，劳动力就多，不但能挑三拣四，价格还贼便宜。
+
+> 都不知道自己基因有多强……
+
+> 一旦65岁以上老年人数量占总人口数的7%，就叫老龄化社会。
+
+> 开个玩笑，大概有以下三个问题
+
+
+# 九、借钱那些事儿：如何避开借贷中的那些坑
+> 裸贷这么不靠谱，怎么这么多人上当？
+
+> 人类的发展史，其实就是个购物史，商品越来越多，就需要在市场上流动起来，但是人们没钱，购买力低，怎么流动起来呢？
+
+> 一手好牌打烂的事儿还少吗？接下来他们通过国家和个人的奇葩遭遇，表达了
+
+> 按行为经济学的研究，最后的结论就是人不理性起来，什么都敢借
+
+> 另外就是看利率，法律就说了，民间借贷的利率高过36%属于高利贷，是非法的。
+
+> 懂点法律很重要，要知道法律是保护咱们的，如果是非法的，就算你入了坑，也可以报警。
+
+
+# 十、消费心理学：消费者防剁手指南
+> 原来大家花钱有很多目的，对于不同的目的，大家愿意花的钱不同。
+
+> 所以，针对低价品，商家喜欢标出折扣；针对高价品就标降价。这两种策略都是为了让你觉得划算。
+
+> 当然，低价品也可以作为锚点。
+
+> 交个订金就好使了？不怕退货吗？
+
+
+# 十一、传销大起底：珍爱生命，远离传销
+> 这个时候大家才发现，这个死胖子根本就没拿钱去投资
+
+> 明白了吧，这整个过程都没有真正的产品，赚钱全靠拆东墙补西墙。
+这个简单的骗术，就是金融界大名鼎鼎的庞氏骗局。
+这个骗局的死穴就在于：需要不断有新钱注入，一旦没有蠢萌新人进来，大家全都要玩儿完。
+
+> 也就是说，传销的本质就是：不要产品，只要人。
+
+> 济学，虽然不太可能成为电视节目里的专家，但能成为小区里的经济学专家，也不错。
+￼
+
+
