+++ 12.微信读书助手/导出资料/古代人的日常生活.txt
@@ -1,0 +1,42 @@
+## 01 古代穿越指南
+> 当人们遥望浩瀚星河，回首千年中华，可能会产生穿越回去的冲动。那么，穿越回哪个朝代会更幸福呢？
+
+
+## 02 古代的厕所长什么样子
+> 春秋时的晋景公，就是如厕时掉进粪坑淹死的。《左传·成公十年》就记载过晋景公“如厕，陷而卒”
+
+> 到了唐朝，由于开国皇帝李渊的爷爷叫李虎，而古人讲究避尊者讳，怎么能把开国皇帝爷爷的名字用在如厕的器具上呢？于是，“虎子”就改名为“马子”。古人对马子的要求也随之变高，不光要能用来装尿，还要能装粪便，所以马子的形制也加大了，变成了桶形，“马桶”因而诞生。
+
+
+## 03 古人上厕所用什么
+> 堂堂国民党高级将领最终栽在了擦屁股用的小小卫生纸上，令人不得不感叹这无常的人生。
+
+
+## 29 古人说话也用文言文吗
+> 古人用文言文的最初目的是为了省字控制成本。
+
+> 宋朝之后的白话文和今天差别不大了，基本上都能看懂。
+
+> 朱元璋有一道圣旨下达抗倭令，要求抵御海盗，原话是这样说的：“奉天承运，皇帝诏曰：告诉百姓们，准备好刀子，这帮家伙来了，杀了再说。钦此。”
+
+
+## 30 古人能听懂我们说话吗
+> 今天的闽南语、粤语、广东客家话和江浙吴语保留了一些古汉语的发音
+
+
+## 48 古人为何瞧不上绿色
+> 古人认为绿色是“苍黄之间”色，即苍天的蓝色和土地的黄色调和而成的颜色，不上不下，颜色不正，是卑微的颜色。
+
+
+## 64 古人有身份证吗
+> 金龟婿”这一说法，就是源于唐朝的鱼符制度。武则天之前，唐朝的鱼符是鲤鱼形的。武则天称帝后，怕人们看到鲤鱼会睹物思人想起“李氏”江山，因为“鲤”和“李”同音，遂将鱼符改成了龟形。龟在古代星宿崇拜中代表玄武，和武则天的“武”谐音。这样，三品以上官员的鱼符就变成了“龟符”，于是就有了“金龟婿”的说法。
+
+
+# 后记
+> 法国历史学家布罗代尔在1958年提出了著名的“历史三段论”，他认为历史应该依据时间分为三类：短时段、中时段、长时段。短时段历史主要指政治事件，帝王将相之事便属于此类；中时段历史则是社会的、经济的、文化的和人口的历史；长时段历史则反映了人们与地理环境的关系，如地理气候、生态环境的变化。短时段是政治的历史，只构成了历史的表面层次，对整个历史进程只起微小的作用；中时段是社会的历史，对历史进程起着直接和重要的作用；长时段是自然的历史，构成整个历史发展的基础。因此，研究历史不应局限于短时段，而应从中时段和长时段去考察，这样才能从根本上把握历史的总体。
+
+> 最后，我选择了读客文化股份有限公司，一是基于对这家公司以往成绩的信任，二是我比较喜欢的历史通俗作家与前辈“二混子”陈磊先生也是这家公司的签约作家。
+
+> 林语堂先生曾经说过：男人的演讲就如同女人的裙子，应该越短越好。这篇后记也应如此。
+
+
