+++ 12.微信读书助手/导出资料/巴菲特给散户的9个忠告.txt
@@ -1,0 +1,16 @@
+# 忠告1 散户要修炼一颗平常心——炒股就是炒心态
+> 要培养自制力，就必须在平时投资时多思考，只有多想想国家的宏观面、股市所处的区域，才能更清晰地认识、理解这一市场
+
+> “投资要有耐心等待，只有等‘市场先生’犯错误，股票被严重低估时才买进。”
+
+> 股价的增长不可能永远超越公司业绩的增长。”
+
+> 其实，股市都有这样一个规律，没有只涨不跌的股票，也没有只跌不涨的股票。涨过了头，就一定要回档进行强制性调整，然后再重新上涨；跌得过多，也必然会调整，随后上升。
+
+
+# 忠告2 投资决策要在能力圈内——不熟不做，不懂不买
+> 巴菲特认为，判断一家公司是否优秀，首先要分析的就是公司的业务。
+
+> 巴菲特认为，越是具有持续竞争优势的企业，其经营业务通常都越简单易懂。投资者在选择投资的企业时，最好选择业务简单易懂的企业。
+
+
