+++ 12.微信读书助手/导出资料/很冷很冷的冷门知识.txt
@@ -1,0 +1,255 @@
+# Question 1 高智商怪咖都关心什么奇葩问题？
+> 但是，人们由此发现，温度对人体的寿命有着极大的影响，人体的衰老正是因为热量在不断地散失。只要把我们的体温降低2℃，这看似微不足道的2℃，就可以让人类多活120年！基于这个“脑洞”理论，“人体冷冻”的科学课题开始走入人们的视线。
+
+
+## 电吹风比X光辐射还要大？
+> 难不成网络上雨后春笋般出现的“脑残”，还真是科技的产物——电吹风制造出来的？
+
+
+## 未来，人类到底能不能去其他星球过日子？
+> “一切皆有可能”
+
+
+## 一滴水的极限杀伤力有多大？千万别跟物理学家较真
+> 俗话说“一方水土养一方人”，
+
+> 如果是液态的一滴水，掉落在地时个头就不会太大，因为空中的阻力会把它撞得“四分五裂”，最终分成几个几毫米直径的小不点。这样的小家伙重量很小，最多不过0.1克，在空中经过漫长的加速之后，最终重力与阻力平衡，速度也就稳定下来，大概保持在10米/秒左右
+
+
+## 用手枪的神枪手都是电影演员
+> 根据大量数据分析，绝大多数使用手枪交战的距离都不超过15米，有信心在15米之外击中目标的，那必须得对自己的水平十分有把握、相当有自信才行——哪怕是经过训练的神枪手也不敢夸下海口。
+
+> 所以，15米几乎可以算是手枪枪战中的“安全黄线”了，要是超过了15米，别怕，就算对方拥有着007的气势，也只能发挥出纯菜鸟的实力，虽然不说100%安全——毕竟运气还是比较玄妙的事情，也能达到99%了。
+
+> 看看武警的训练要求，使用手枪的时候，他们有特殊的标准，比如3～10米的距离内，能够在3秒钟之内连续射出3发子弹，并且达到至少一次击中敌人的目标就算合格。是的，你没有看错，不是100米，不是30米，只要3米！难道他们中没有神枪手？不说比得上奥运冠军，至少也比电影里的小混混强得多。
+
+
+## 电脑里的A、B盘去哪儿了？
+> 这事要是问售后，售后还真是一个大写的“冤”——就算藏了，也不是售后干的呀。咳咳，当然，这也跟售前没关系，更不是厂家侵吞了您的利益，电脑的硬盘还真就是从C盘开始算的。
+这么奇怪的标记方法，难道是发明人脑子一时糊涂，把A、B盘给忘了？还真不是！A、B盘在一开始，的确是我们电脑不可或缺的一部分，不过它们不属于硬盘系统，而是最开始家用电脑普及的大功臣——软盘。
+
+> 硬盘的成本很高，价格更是令人咂舌。商用、研究用的电脑上安装硬盘没问题，反正公司有钱
+
+> 一开始，软盘的尺寸有8英寸，相当于一个iPad迷你——这可一点都不迷你，后来缩小到5.25英寸，终于跟电脑“适配成功”，成为家用电脑的第一个存储设备——A盘。软盘虽然便宜，但是存储量很小，放到现在也就能够存储一首歌，还不能是高保真的，因此后来就设计为“双驱动”，也就是插两张软盘。这样，B盘就诞生了。
+
+> 现在，我们的电脑上已经再也看不到插入软盘的接口了，似乎软盘的痕迹已经被清扫干净。不过，显然人们不想对这样的大功臣“利用完就扔”，因此还是保留了软盘的位置，那就是电脑的A、B盘。
+哪怕软盘已经被淘汰，命名的规则也没有改变，消失不见的A、B盘，反而让我们能一直记住这个家用电脑推广的大功臣呢！
+
+
+## 拆定时炸弹为何要剪红蓝电线？
+> 其实仿照的都是最简单的定时炸弹，拥有两支线路的并联电路。
+
+> 当两条支路并联在电路上的时候，电流会全部流经没有电阻的支路，这就是我们所说的“短路”。所以，装有起爆器的电线与一根空载的电线并联，电流会完全忽略起爆器这个“大杀器”，而是全部绕行，使其短路。
+
+
+# Question 2 历史“遗留”问题研究中心
+> 除了麝香，水银也是古代十分恐怖的避孕药，它多用于青楼女子身上。
+
+> 元朝人孔齐就发现，人们可以通过算日子的方式避孕。
+
+> 劳动人民的智慧是无穷的，他们从动物身上发现了简易版的“避孕套”，比如猪膀胱、羊肠、鱼鳔……还别说，这些东西轻薄又有弹性，与现代的“杜蕾斯”们的确有异曲同工之妙。
+
+> 江浙的男人想要避孕的时候，往往会吃一种叫“棉籽油”的食物，吃的时候不会使人怀孕，停上一阵子就又能恢复了
+
+
+## 古代处决犯人为什么安排在秋季？
+> 为什么古代只在秋天处决犯人？
+这还不简单，没听说过成语“多事之秋”吗？
+这说明，秋天就是一个忙碌的季节，刽子手也该忙起来了！
+
+> 春天庆祝、夏天赏赐、秋天处刑、冬天行罚，这样就是顺天而行。
+
+> 古今中外人们似乎都偏爱在大庭广众之下行刑，其实就是为了“杀鸡儆猴”，让围观的人知道违法乱纪的后果，以警示他们。要是选择春夏两季行刑，农民们都忙着种地，秀才们忙着准备科举考试，谁没事出门乱逛来观看行刑呢？而秋冬时节则是大多数人空闲的时候，行刑的效果更好。
+
+
+## 古人大便后用什么擦屁股？
+> 奢侈些的如英国王室，在15世纪前使用鲑鱼擦屁股
+
+
+## “出门饺子进门面”源于什么脑洞？
+> 出门饺子进门面”源于什么脑洞？
+出门要吃饺子，
+
+> 习俗！习俗就是那些你永远理解不了为什么的东西。
+
+
+## 为什么是“上厕所”与“下厨房”？
+> “上厕所”“下厨房”这两个词怎么来的，你想过吗？
+难道古代的厕所都在楼上、厨房都在地下室？
+
+> 东边的厢房一般住家里的“少爷”，西边则是一众“小姐”。
+
+> 人们认为北方利水，东方利木，而水跟木是相生的，厕所建在这里，意味着有水——你就不用去联想是什么水了，可以生木，对耕种有好处。
+
+
+## 古人上班打卡吗？作息时间是怎么定的？
+> 知道这时候太阳“上班”晚，就将上班时间推迟到了早上六点。没错，六点！那要是适合早起的夏秋两季呢？五点半，不好意思，不能迟到哦！
+
+> “点卯”这个词就是这么来的，因为点名签到的时间都规定在卯时，所以如此。卯时是现在的几点呢？早上五点到七点
+
+
+# Question 3 那些令人肾上腺素飙升的问题
+> 你听说过这样的流言吗？
+某某村的某某，被偷走了两个肾！听说现在还在抢救呢！
+不是吧？这么危险？不过……还抢救什么呀？
+为什么这么问？
+两个肾全没了，人还能活？
+
+> 都说到这个程度了，相信你也听出来了，关于偷肾的那些流言，99%都是假话，剩下的1%也可能是误会。而且，其中至少有一半以上，还写着“转发起来提醒大家”“不转不是中国人”等字眼。为什么？骗你转发呗！
+
+> 在中国流传最广泛的版本，是一个可怜的四川姑娘。这个涉世未深的大学生，因为和网友开房，昏迷在了宾馆中。醒来后发现自己躺在装满水的浴缸里，身上剧痛难忍，腹部有两道伤口。她的身边有个纸条，上面写着“快打120，否则你会没命的”
+
+> 这说明了什么？梗不在老，有趣就行。世界人民都爱八卦，一个“卖肾”的故事，一样能传唱二三十年，还不断推陈出新。
+
+
+## 葬礼上为什么要戴黑纱？
+> 在中国传统的丧礼中，似乎并没有见过戴黑纱这个环节。中国以白色作为哀悼的颜色，因此传统的葬礼上，亲人往往都穿白衣、头扎白色布条。所以丧事又被叫作“白事”，这一身穿戴又被称为“披麻戴孝”，是孝子贤孙送葬的必备制服。
+
+> 一些仆从囊中羞涩，买不起黑色丧服，又不能不追悼主人，就采取了简化方式——在左臂上佩戴一块黑纱
+
+
+## 为什么我们会害怕蛇和蜘蛛？
+> 所以，这种害怕往往是没有原因的，也许蛇和蜘蛛也很纳闷呢。
+
+> 可以说，就是因为我们的老祖宗把对它们的恐惧与厌恶写在了基因里，才会让我们变成敏感的胆小鬼。不要因此嘲笑我们祖先胆小，这样的基因才是在恶劣的远古时代成功存活的“保命利器”，是进化的标志。
+
+
+## 为什么我们会有密集恐惧症？
+> 一个个生活中看似正常的家伙高呼自己“密恐”了，好像不“密恐”一下，就要被人踢出时尚圈一样。
+
+> 他们害怕的事物有一个共同点，就是颜色对比比较强烈，并且在空间内通过特定的规律排列，同时十分密集。
+
+
+## 看恐怖片真的可以减肥吗？
+> 根据英国《每日邮报》的一篇文章介绍，科学家发现当我们观看恐怖片的时候，会因为恐惧而出现心跳加快、代谢增速的状况，从而消耗大量的热量，达到减肥的目的。先不考虑看恐怖片的减肥法能不能顺利实施，理论上，这的确是可行的。
+
+
+# Question 4 非正常人类最关注问题集锦
+> 关于火车上的排泄物去哪了，相信大家只要坐过车、观察过就有结论，它们会落在沿途的铁轨边，最终化作天然粪肥。
+
+> 飞机处理排泄物的方式，的确跟火车有着异曲同工之妙，那就是——在半途直接丢掉。
+
+> 为了防止无辜路人再次中枪，人们锲而不舍地研究飞机上的马桶技术，最终出现了现在常用的马桶系统——真空马桶。这种马桶相当于是老式马桶+吸尘器的强强联合，里面藏着一个神奇的吸尘器，可以先把粪便直接“吸走”，然后再用少量的清水冲洗内壁即可。忘了说，连马桶的内壁也是特制的，别看它一辈子都只能跟粪便接触，用的却是高级的不粘材料，保证不会有零星的“金坷垃”粘在上面，有碍观瞻。
+
+
+## 打哈欠真的能够传染
+> 一项研究表明，打哈欠其实就是在给我们的大脑降温。他们认为，我们人类的身体跟一台电脑很相似，心脏就是电池，而不断运转的CPU就由大脑代替。CPU负荷过大的时候电脑就会发热，此时要是没有风扇吹着降温，电脑分分钟能够给你闹罢工。而大脑也是如此，如果温度过高，我们就会感觉疲惫，也就是俗称的“脑子不转了”。此时，打哈欠就成了天然风扇，可以让我们的大脑清醒。
+
+> 这种魔性的传染行为，只在人类、黑猩猩和狗等生物身上发现过。也就是说，想要跟风打哈欠，还必须得有高智商，只有大脑皮层发达、能够辨别出打哈欠的动作，才能传染。用这一个简单的动作，就可以将动物们按照智商分级了，实在非常容易。
+
+> 由此延伸，你就会发现，那些社交能力较好、敏感热情、相对感性的人，更容易在别人打哈欠的时候被传染；相反，理性的、性格冷酷的家伙，就比较能坚持自我。所以，通过能不能被感染，还能看出我们性格中的一些特点。
+
+> 如果你想看看自己跟身边人的熟悉程度，不如就找个机会用打哈欠来证明一下吧。
+
+
+## 怎样证明你不是一个神经病？
+> 心理学家罗森汉突然想调戏一下精神病医生，他想知道，如果找人装成精神病，医生能不能发现呢？
+
+> 如此一来，这41个病人到底是不是精神病呢？想一想，还真是细思恐极啊！
+
+
+## 鸡走路为什么总是颠脖子？
+> 颠着脖子咯咯叫，它不像白鹅一样迈着优雅的“四方步”，而是伴随着颠簸的小碎步，脖子不停地一伸一缩，看起来很有意思，这也让我想起当年边听周杰伦边自我陶醉的样子。
+
+> 它们并不是让脑袋前后伸缩，而是先伸出脑袋，静止在那里等待着身体和脚跟上。也就是说，鸽子和鸡的脑袋是个“急性子”，移动的时候总是快身体一步，所以只好安静地在原地等着身体的其他部位。
+都说思维太快，行动都跟不上了，原来是这个意思。
+
+
+## 为什么喜欢的人要“么么哒”？
+> “我们在找到王子的过程中，可能要亲吻无数只青蛙
+
+
+## 日本女生的校服为什么是水手服？
+> 大概人们都有一种军人情结，对制服有着特殊的喜爱，
+
+> 后来，日本人从中发现了好处——水手服作为制服，体现的是一种军事化的含义。当时，人们满脑子都是“富国强兵”的念头，他们显然也很懂“要从娃娃抓起”的概念，两者一结合，最简单快捷的办法就是换校服。
+
+
+# Question 6 如何证明你是一个文化人？
+> 衰老的水还有一个特殊的名字，叫“老化水”
+
+> 哪怕是纯净水，放上一天也能出现0.0004毫克/升的亚硝酸盐，只要放上20天，其含量就能追赶到0.73毫克/升了。这大自然的魔术，实在让人防不胜防。
+
+> 如果是未成年人经常饮用这种老化水，很容易影响生长发育，细胞的代谢速度也会减缓，而中老年人更明显，还可能出现消化神经等系统的问题，导致早衰。
+
+
+## 故宫门上的门钉有什么含义？
+> 吸引着无数人咽下这块有毒的蛋糕。
+
+> 在清朝，门钉的钉法、数量皆有规定，要是谁家不小心多钉了几个，很容易被认为是“藐视皇权”，抓进大牢没商量。
+
+
+## 为什么苏格兰人爱穿裙子？
+> 在穿裙子方面，苏格兰男性绝对堪称教师级别。他们对裙子的热爱完全不亚于女性，事实上，你让一个姑娘脱下裙子换上裤装，她们可能还欣然接受，让一个苏格兰男人从此不穿格子裙，不好意思，他能立刻跟你拼命
+
+> 从几年前的苏格兰独立公投就能看出，哪怕是到了现在，倔强的苏格兰人还是不愿意向南方的英格兰人低头。
+
+> 表面上，这是苏格兰人“爱裙子”的特殊文化，实际上，这代表了苏格兰民族几百年来不服输的抗争。而这一切，都蕴含在一条小小的短裙中。
+
+
+## 啤酒瓶盖为什么是21个齿？
+> 现在我们能看到的啤酒瓶盖，那都是按照国际标准生产的，仔细数一数，全部都是21个齿。
+
+> 目前通行的啤酒瓶盖，全部是根据德国制定的标准“DIN6099”制造出的。
+
+
+## 为什么美国人把足球叫Soccer？
+> “soccer”这个词其实由来已久，甚至比现代足球的历史还要长。早在欧洲中世纪的时候，英国就出现了“soccer”这项运动，它也的确是一项跟脚、球有关的运动，但并不是现代的足球。
+
+> 直到二战后的一段特殊时期，英国人不知道被触碰了哪根神经，突然开始讨厌自己的美国兄弟了。就好像较劲一样，美国人用什么，英国人偏偏就不想用。比如美国口音的英语，在他们眼中就是“俗”的代表，美国人的生活习惯，也被英国“绅士”们所鄙夷。总之，完全是一副两兄弟分家产、撕破脸的状态。
+
+
+## 飞机遇险时为何不能跳伞逃生？
+> 每当看到飞机发生空难的新闻时，相信不少人都有疑惑：这都千钧一发的时刻了，为什么没有人选择跳伞呢？可别说一飞机的人全都恐高啊！
+
+> 正常飞行的时候，飞机的高度一般都在8000-10000米以上。这是什么概念？窗外的气压低到几乎打不开飞机门，要是把机舱打开，乘客别说站在飞机边准备跳伞了，恐怕分分钟就会连着座椅一起被吸出窗外。
+
+> 而飞机最容易遇到危险的时候，恰好又是在起飞、降落的阶段，这个高度的环境倒是适宜人生存的，可惜——太低了。乘客们跳出窗外，可能还没反应过来要开降落伞，就“吧唧”一声酿成惨剧了。就算打开了降落伞，没有足够的高度消除冲力，一样很容易导致受伤。所以，在低空中跳伞也不成立。
+
+
+# Question 7 不是你不明白，是这世界变化快
+> 真不知道，这群冰球观众们到底是来看球的，还是来看打架的。其实，就像斗牛、拳击等运动备受喜爱一样，人们似乎天生对与“打架”有关的运动格外感兴趣，自己不能亲自上阵，也得看别人热血沸腾一把。
+
+
+## 国产电视剧胸口中弹为何总是七窍流血？
+> 表演的精髓就在这里，我痛不痛苦不重要，重要的是要让观众感受到痛苦。
+
+> 一般胸部中弹是不会导致口鼻流血的。子弹射中了胸腔，穿透的往往是肺部、心血管、心脏等。如果是心脏不幸中弹了，血液肯定是从伤口处往外喷射，这里失血的速度就够快了，根本赶不及再从口鼻流出去。如果是肺部中弹，最明显的后果就是呼吸困难、缺氧，胸部可能会出现大量的血泡，也不会有口鼻流血的状况。
+
+
+## 听说日本人的姓氏很奇怪？
+> 也就是说，日本人有姓氏的历史，其实才一百多年。
+
+> 比如“十七女十四男”这样的姓，难道不是他们在吐槽对方家庭人口过于繁盛吗？
+
+> 有趣的是，虽然全日本的人现在都有了姓氏，天皇却是没有姓的。这是因为他们认为，天皇不属于“人”的范畴，不需要有家族、姓氏的束缚。
+
+> 真的不是因为天皇嫌弃这些姓太难听，所以才拒绝的吗？
+
+
+## 对号为什么写成“√”？
+> 先来看看它在西方的历史。“√”的起源来自英国，一开始的意思就是“对”“正确”。和这个国家糟糕的中小学数学教育不同的是，17世纪英国的数学家层出不穷，当然现在也不少——虽然我们很疑惑，这么差劲的数学教育是怎么培养出数学家们的。
+
+> 由于“r”这个字母写起来特别像“√”，而后者更加简单方便，于是它就在教师圈子里流行起来了。
+
+> 明朝时，狡猾的严嵩同志就是利用这一点，在给皇上呈名单的时候偷偷加上了眼中钉杨继盛的名字，结果粗心的嘉靖帝没仔细看给打了个钩，一代直臣、明朝第一硬汉就这么冤屈被斩了。
+
+> 在中国古代，人们的确是频繁地使用“√”，而且它往往是“已经检查，正确无误”的意思，也有“对号”的作用。这么看，也许真的是中英两国人民的脑洞开到一起去了呢！
+
+
+## 为什么化妆品专柜总是在一楼？
+> 没错，商场中将化妆品柜台规划在一楼，归根结底就是为了让商场人流量更多，吸引和促进人们消费。
+
+
+## 为什么衬衫上的扣子是“女左男右”？
+> 女式衬衫的扣子都在左边，而男式衬衫的扣子就在右边
+
+> 一种说法认为，中世纪的男性大多数都是“战斗狂人”，随时随地都能跟对面的绅士脱手套拔剑来一场决斗，所以右手往往是留着用来持剑的。此时，空出来的只有左手，为了方便男人们帅气地用单手扣扣子，就把纽扣设计在了右边。
+
+> 哪怕上面一粒小小的扣子也有有趣的历史，哪怕这段历史已经模糊了，它也依旧能在我们的生活中发光发热。
+
+
+## 正在吸血的蚊子，只能留它一命？
+> 吸血的蚊子如果被打死，口器就很容易残留在皮肤里，这样皮肤上就会留下肉眼难辨的小伤口难以愈合。
+
+
