+++ 12.微信读书助手/导出资料/新型冠状病毒感染防护.txt
@@ -1,0 +1,4 @@
+## 四、针对新型冠状病毒，该如何进行消毒？
+> 新型冠状病毒怕热，在56℃条件下，30分钟就能杀灭病毒；含氯消毒剂、酒精、碘类、过氧化物类等多种消毒剂也可杀灭该病毒。
+
+
