+++ 12.微信读书助手/导出资料/最冷最冷的冷门知识.txt
@@ -1,0 +1,127 @@
+# 丰富渊博的历史知识
+> 当然，古人也并不都是使用竹片。有的地方，石头、鹅卵石等都可以用来当作擦秽之物。不仅洗一洗还可以再用，而且还环保。到了唐宋时期，粗纸已经开始被使用，一直到元朝，纸才被广泛运用于入厕。我们知道，汉代就发明了纸，但为何这个改变人类文明进程的东西直到元代才被国人用来擦屁股呢？后人揣测，元朝是蒙古人建立的，文化相对比较落后，没有汉民族“敬惜字纸”的意识，所以才使得厕纸进入人们的生活。
+
+> 蔡邕作为议郎，曾被汉灵帝召进宫中商策。他直言不讳地斥责“帝贪，政暴，吏酷”
+
+> 蔡邕就这样绊绊磕磕、慌慌张张、窝窝囊囊地活了一阵子。
+
+> 可惜的是，是否要判蔡邕死刑的争议尚未停止，蔡邕就死在狱中了
+
+> 1900年6月21日，农历五月廿五，当时中国的实际统治者慈禧太后做了一件空前绝后的大事——向全世界宣战。慈禧在《诏书》中信誓旦旦地说：“与其苟且图存，贻羞万古；孰若大张挞伐，一决雌雄。”如此气势，完全不是那个丧权辱国的老佛爷作风。
+
+> 锦衣卫，全称“锦衣亲军都指挥使司”，原为明朝内廷御林军，后成为皇帝的军政特务机构。
+
+> 如今，我们一般把刀、枪、剑、戟、斧、钺、铲、叉、鞭、锏、锤、戈、镋、棍、槊、棒、钯、矛等称为十八般兵器。
+
+> 瞒天过海、围魏救赵、借刀杀人、以逸代劳、趁火打劫、声东击西、无中生有、暗度陈仓、隔岸观火、笑里藏刀、李代桃僵、顺手牵羊、打草惊蛇、借尸还魂、调虎离山、欲擒故纵、抛砖引玉、擒贼擒王、釜底抽薪、浑水摸鱼、金蝉脱壳、关门捉贼、远交近攻、假道伐虢、偷梁换柱、指桑骂槐、上屋抽梯、树上开花、反客为主、假痴不癫、美人计、空城计、反间计、苦肉计、连环计、走为上。
+
+> 33.唐伯虎未曾点秋香
+
+> 历史上的唐伯虎也许根本就没有“点秋香”的艳遇。
+
+> “衣冠禽兽”一语最早来源于明代官员的服饰。明朝规定，文官官服绣禽，武官官服绘兽。品级不同，所绣的禽和兽也有所不同。 
+    具体的规定是：文官一品绣仙鹤，二品绣锦鸡，三品绣孔雀，四品绣云雁，五品绣白鹇，六品绣鹭鸶，七品绣鸳鸯，八品绣黄鹂，九品绣鹌鹑。 
+    武官一品、二品绘狮子，三品绘虎，四品绘豹，五品绘熊，六品、七品绘彪，八品绘犀牛，九品绘海马。
+
+> 在服饰的颜色上，不同的品级，也有不同的颜色。文武官员一品至四品穿红袍，五品至七品穿青袍，八品和九品穿绿袍。所以，当时的“衣冠禽兽”指的是官员的服饰，有令人羡慕的味道，一般用作赞语。
+
+> 日常生活中，人们常把傻瓜或处事随便、说话不正经、好出洋相的人叫作“二百五”。这种说法来源于战国时期，跟苏秦有关。
+
+> 后来在齐国被人暗杀。对此，齐王很恼怒，发誓要为苏秦报仇
+
+> “二百五”一词就这样流传了下来，从这个词的来源看，用来形容那些傻乎乎、好出洋相的人，倒也贴切
+
+
+# 不得不知的国学常识
+> “中国”一词有3000年的使用历史，最早出现于周朝。当时因为华夏民族拥有了相对先进的农耕文明，并且建立了一套完善的礼仪制度，而周围的四夷仍旧裹着树叶兽皮靠打猎为生，于是把自己与四夷分开，以显示自己的优越
+
+> 《大学》一书，深刻地阐述了修身、治国的道理和方法。
+
+> 八卦是什么
+
+> 鬼谷子，姓王，名禅，字诩，战国时期卫国人，著名的思想家、谋略家、兵家、教育家，是纵横家的鼻祖，在中国历史上极富神秘色彩。据说其“鬼谷”之名，是由其出生地或隐居地得来。 
+    鬼谷子是一位全才型的人物，他既有政治家的六韬三略，又擅长外交家的纵横之术，更兼有阴阳家的祖宗衣钵，预言家的江湖神算，被人们看作一位奇才。鬼谷子还是一位伟大的教育家，他比较知名的兵家弟子有孙膑、庞涓，纵横家弟子有苏秦、张仪，这四人都是战国时期的风云人物。
+
+> 梦溪笔谈
+
+> 襁褓——不满周岁； 
+    黄口——本指雏鸟，后比喻10岁以下的儿童； 
+    总角、孩提——幼年的泛称； 
+    垂髫（tiáo）、始龀（chèn）——童年的泛称； 
+    幼学——10岁； 
+    束发——15岁左右； 
+    弱冠——20岁； 
+    而立之年——30岁； 
+    不惑之年——40岁； 
+    知命之年、年逾半百、知非之年、艾服之年、大衍之年——50岁； 
+    花甲、平头甲子、耳顺之年、杖乡之年——60岁； 
+    古稀、杖国之年、致事之年、致政之年——70岁； 
+    杖朝之年——80岁； 
+    耄耋（mào dié）之年——八九十岁的年纪； 
+    鲐背之年——90岁； 
+    期颐（qī yí）之年——100岁。
+
+> 原来，古人为了防止书籍被蠹虫损坏，一般会在书籍里面夹上几片香草。这种香草叫芸草，是多年生草本植物，有特异的清香气，即使枯萎香气也不会变淡，它可以驱除蠹虫，古诗中就有“芸叶熏香走蠹鱼”的描写。所以“书香”的“香”指的是“芸草香”。
+
+
+# 舌尖上的饮食文化
+> 张骞出使西域之后，从西域引进了很多植物，如茄子、大蒜、西瓜、黄瓜、扁豆、刀豆等，极大地丰富了供人食用的蔬菜种类
+
+> 彭祖是我国传说中最长寿的人，他的传奇为后世很多人津津乐道，包括孔子和庄子。
+
+
+# 千奇百怪的生活知识
+> 第一个原因是：我们的身体皮肤非常敏感，只要身体的皮肤一碰到热水的时候，血液循环就会变快，人的膀胱就会变得膨胀，所以就会感到有尿意。
+
+> 而用热水灭火，热水能在短时间内迅速汽化，产生水蒸气，把火源覆盖，使火源得不到充足的氧气而无法燃烧，所以，热水能够从两个方面来遏制火势的发展。
+
+> 毛主席说“屁乃五谷杂粮之精华”
+
+> 芦荟是含有毒素的植物之一，中毒剂量为9～15克，通常情况下，在食用后8～12个小时内会出现恶心、呕吐、腹泻、剧烈腹痛、出血性胃炎等症状。
+
+> 在西方国家，人们认为人死后都是有灵魂的。当人们在参加丧礼的时候，死者的灵魂会在天上看着他们，这个时候，如果你穿着颜色太过鲜艳的衣服，就会被死者的灵魂看到，然后就会被死者的灵魂残害致死。但是，如果穿黑衣服的话，就不会那么容易被死者的灵魂发现了。身穿黑色的丧服其实不是为了哀悼死者，而是惧怕死者的灵魂。
+
+> 所以，请一定要记清楚，车祸跳车逃命时，一定要选择向后跳哦！
+
+> 纽扣为什么会“男左女右”呢
+
+> 连如今男女结婚拍结婚照也是“男左女右
+
+> 饭后饮茶是一个非常不好的习惯。这是因为：饭后，胃部受到食物的刺激，会分泌大量胃溶液来溶解食物。食物被溶解后，才能被我们的消化系统所吸收。如果我们在饭后立即喝茶的话，茶水会冲淡胃部所分泌出来的胃液，减慢消化。同时，茶叶里还含有一种化学成份，可以使食物凝固而影响人体对蛋白质的吸收。
+
+> 切洋葱的时候，由于洋葱散发出的味道刺激性比较大，往往把我们弄得泪流满面。其实，在我们切洋葱的时候，只需要在嘴里嚼片口香糖就可以轻松化解流泪的难题。
+
+> 最简单的方法就是拿着化妆水使劲儿摇，摇完后看泡泡： 
+    （1）泡泡少，说明营养成分少； 
+    （2）泡泡多但大，说明含有水杨酸，洁肤的效果较好，但刺激性大，易过敏； 
+    （3）泡泡很多很细，而且很快就消失了，说明含酒精，不要长期使用，容易伤害皮肤的保护膜； 
+    （4）泡泡细腻丰富，有厚厚的一层，而且经久不消，那就是好的化妆水。
+
+> 喝酒怎么做才不容易醉
+
+> 所以，深受压力之害导致失眠的人们，千万别再尝试“数羊”了，这样只会让你更加失眠。
+
+> 34.本命年为什么要穿红色的内衣
+
+> 一个男人，终其一生所追求的不外乎钱、权和性，而钱和权在很大程度上也是为了满足自己的性欲，所以不夸张地说，男人一生最大的追求就是性。只有在性上满足了，男人才能真正获得快乐，才会意气风发；如果得不到满足，男人就会异常压抑。
+
+
+# 五花八门的科学知识
+> 当人体的温度低于30摄氏度的时候人就会被冻死。
+
+> 3.狗为什么不能吃巧克力
+
+> 打火机发明于16世纪，而火柴发明于19世纪。
+
+> 在单位时间内，人迎面遇到的雨水为：V2=vts·l·t 
+    所以在雨水的各性质相同时，跑步与慢走相比，迎面遇到的雨水较多，而上面遇到的雨水相同，因此跑步的人淋
+
+> 我们在打哈欠时张口猛吸一口气，肺部就会扩张得比平时大，呼气时，肺部又能尽量地收缩，把二氧化碳排出去了，这样体内的氧气就能很快补足了，人很快就会为之精神一振。
+
+> 开灯睡觉的人或者生物钟自然睡眠模式受人造光线干扰的人，患癌症的可能性比平常人要大很多。
+
+
+# 丰富多彩的生物世界
+> 在地球上，有一种神奇的树木，叫作梓柯树，这种树是不怕被火烧的。
+
+
