+++ 12.微信读书助手/导出资料/请回答1998.txt
@@ -1,0 +1,43 @@
+# 版权信息
+> 宿舍里的每个人
+都给父母打电话，问平安。
+
+
+## 一、请回答1998
+> 一开始，我还正襟危坐在小板凳上，后来离电视机越来越近，最后整个人都站起来，脸要贴在屏幕上，被回家的父亲批评：“这样看电视眼睛会坏掉的！以后还这么近，就把电视卖了，全家督促你学习，省得你天天偷看。”
+
+
+## 二、1999，全世界伤心角色又多我一个
+> 父亲工作很忙，但过年打扫房屋院落，是男人们应该做的。
+
+> ，练习《走进新时代》这首曲子。
+我只要在家里一弹这首歌，父亲总会跟着唱几句：“总想对你表白，我的心情是多么豪迈，总想对你倾诉，我对生活是多么热爱，勤劳勇敢的中国人，意气风发走进新时代，我们意气风发，走进那新时代。”
+
+> 我把炭盆装得满满的，像装了一盆珍宝，快到家门口的时候，大喊：“妈，快点开门，我坚持不住了。”母亲匆匆把门打开，掀起门帘，看我端着满满一盆炭，脸憋得红红的，说：“啊呀呀呀，你这孩子是不是傻，端上这么多干什么！”
+
+> 晚上看春晚，我最喜欢的相声演员冯巩和郭冬临一起合作了小品《旧曲新歌》。
+
+> 家乡过年，春联、字楼、柏枝、福字、灯笼，一个都不能少。
+
+> 男人们聊起这一年的收入，挣钱的门道，明年的打算，女人们照顾着孩子吃饭，互相问起孩子上几年级了，是否爱学习，又扯到邻居家的八卦，村里的婚丧嫁娶。
+
+
+## 四、出生在中秋的孩子
+> 1992年的中秋节，我出生了。六年后的中秋节，我弟出生。
+
+> 父亲来了兴致，要我和弟弟来一段现场演出，我们俩两眼相望，只好做足戏，来个激情洋溢的自我介绍，以及滥竽充数的绕口令，“扁担长，板凳宽，扁担绑在板凳上……”
+
+> 那时候，等吃了午饭，母亲总会领着我和弟弟去照相馆的赵爷爷那里，拍一张照片留作纪念。后来我一张张翻过，时间真的会在一个人的脸庞、身体上留下痕迹。可惜，那些照片里只有我和弟弟果断、暴风成长的景象，却没有父母的影子。
+如果有的话，也只怕让人看着心酸。
+
+
+## 五、父母爱情
+> 每当一束相似的灯光出现，母亲就探着身子，偏颈张望，又不时地低头拽紧我，语气迟疑，“这个不敢定是你爸。”等它走近，母亲又说：“不是啊，再等等吧。”这样失望和欣喜交织的情绪，会不停地、反复地出现，在这个夜里。
+
+> 父亲和母亲很少吵架。小时候，每当他们俩吵架，我就会想很多，最令我头疼的问题是：如果他们俩离婚，我要跟谁？
+
+
+## 六、戏台
+> 他讲着讲着，我的眼睛就开始发困，于是趴在他膝头睡着了。外公看见我睡着，也不再说什么。他静静地在院子里坐一会儿，也不着急带我回去，反而轻拍我的后背，任由我呼呼睡着，像是要留住这短暂的时光。
+
+
