+++ 12.微信读书助手/导出资料/这是你的星星，请签收.txt
@@ -1,0 +1,23 @@
+# CHAPTER 1 遗失
+> 夜空中的星星越多，美梦就越多。
+
+> 小饿魔负责把所有的噩梦统统吃掉，因此胖胖的。
+
+
+# CHAPTER 3 找寻
+> 星星告诉我自己做的选择必须由自己承担后果，所以我不知道还可以回去。
+
+> 看清楚刚刚面前的一堵墙原来是两棵巨大的仙人掌。此刻小刺猬正站在他们中间一脸幸福地伸出手介绍道：“喂，茶仔，这是我的爸爸，这是我的妈妈。”
+
+> “你也一定会找到你爱的人，因为你们始终住在彼此心里。”
+
+
+# CHAPTER 5 南极的星星
+> 其间，有另外一只企鹅每天按时给企鹅姑娘送来各种新鲜的鱼吃，
+然后默默走开。
+
+
+# CHAPTER 6 任意门
+> “老爷爷或许也经历过地铁里遇到的年轻人那样的焦虑与迷茫，老奶奶呢，或许也像小企鹅一样曾遭受失恋的重击吧。”
+
+
