+++ 12.微信读书助手/导出资料/韭菜的自我修养.txt
@@ -1,0 +1,15 @@
+# 前言
+> 对于那些“习惯性忽略大面积文字”，且“习惯性自由组织局部内容”的阅读障碍症患者来说，太容易产生诡异的歧义了。
+
+
+# 3 亡羊补牢的“韭菜”才有前途
+> 所以，“亡羊补牢”是有前提的：
+ 
+●你有的不仅仅是亡掉的那一只羊；
+●除了它之外你最好还有很多只羊；
+●因为你有很多只羊，所以你过去就建过“羊牢”；
+●所以，你虽然“亡羊”了，但还有机会去“补牢”；
+●所以，一旦“牢”补好了，你可以管更多只羊；
+●所以，最初的时候亡掉一两只羊，对最终结果的影响可忽略不计。
+
+
